backend/requirements.txt
fastapi==0.110.3
uvicorn==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.6.1
python-dateutil==2.9.0.post0
email-validator==2.2.0
firebase-admin==6.5.0
google-generativeai==0.8.5
hindsight-all
